requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
watchdog==3.0.0

# Document processing
pymupdf==1.23.21
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
python-multipart==0.0.7

# Vector database
qdrant-client==1.7.0

# Machine learning
transformers>=4.36.0
sentence-transformers
tqdm
requests>=2.31.0
accelerate
rank_bm25==0.2.2
nltk==3.8.1
openai>=1.3.0
tiktoken>=0.7.0
spacy>=3.7.3

# Install PyTorch separately with:
# pip install torch --index-url https://download.pytorch.org/whl/cpu
